have pretty much all functionality for APIs working as it should

+++ ApiDoc/Models/VersionedItem.cs
@@ -6,7 +6,7 @@ namespace ApiDoc.Models
     {
         public string Author { get; set; }
         public DateTime ChangeDate { get; set; }
-        public long RevisionNumber { get; set; }
+        public int RevisionNumber { get; set; }
         public bool Deleted { get; set; }
     }
 }
